Upgraded core files to version 1.3.8

+++ cake/tests/lib/cake_test_fixture.php
@@ -4,14 +4,14 @@
  *
  * PHP versions 4 and 5
  *
- * CakePHP(tm) Tests <https://trac.cakephp.org/wiki/Developement/TestSuite>
+ * CakePHP(tm) Tests <http://book.cakephp.org/view/1196/Testing>
  * Copyright 2005-2010, Cake Software Foundation, Inc. (http://cakefoundation.org)
  *
  *  Licensed under The Open Group Test Suite License
  *  Redistributions of files must retain the above copyright notice.
  *
  * @copyright     Copyright 2005-2010, Cake Software Foundation, Inc. (http://cakefoundation.org)
- * @link          https://trac.cakephp.org/wiki/Developement/TestSuite CakePHP(tm) Tests
+ * @link          http://book.cakephp.org/view/1196/Testing CakePHP(tm) Tests
  * @package       cake
  * @subpackage    cake.cake.tests.libs
  * @since         CakePHP(tm) v 1.2.0.4667
@@ -68,7 +68,10 @@ class CakeTestFixture extends Object {
  */
 	function init() {
 		if (isset($this->import) && (is_string($this->import) || is_array($this->import))) {
-			$import = array_merge(array('connection' => 'default', 'records' => false), is_array($this->import) ? $this->import : array('model' => $this->import));
+			$import = array_merge(
+				array('connection' => 'default', 'records' => false), 
+				is_array($this->import) ? $this->import : array('model' => $this->import)
+			);
 
 			if (isset($import['model']) && App::import('Model', $import['model'])) {
 				ClassRegistry::config(array('ds' => $import['connection']));
@@ -77,6 +80,7 @@ class CakeTestFixture extends Object {
 				$db->cacheSources = false;
 				$this->fields = $model->schema(true);
 				$this->fields[$model->primaryKey]['key'] = 'primary';
+				$this->table = $db->fullTableName($model, false);
 				ClassRegistry::config(array('ds' => 'test_suite'));
 				ClassRegistry::flush();
 			} elseif (isset($import['table'])) {
@@ -88,23 +92,24 @@ class CakeTestFixture extends Object {
 				$model->table = $import['table'];
 				$model->tablePrefix = $db->config['prefix'];
 				$this->fields = $model->schema(true);
+				ClassRegistry::flush();
+			}
+
+			if (!empty($db->config['prefix']) && strpos($this->table, $db->config['prefix']) === 0) {
+				$this->table = str_replace($db->config['prefix'], '', $this->table);
 			}
 
 			if (isset($import['records']) && $import['records'] !== false && isset($model) && isset($db)) {
 				$this->records = array();
 				$query = array(
-					'fields' => array_keys($this->fields),
-					'table' => $db->fullTableName($model->table),
+					'fields' => $db->fields($model, null, array_keys($this->fields)),
+					'table' => $db->fullTableName($model),
 					'alias' => $model->alias,
 					'conditions' => array(),
 					'order' => null,
 					'limit' => null,
 					'group' => null
 				);
-
-				foreach ($query['fields'] as $index => $field) {
-					$query['fields'][$index] = $db->name($query['alias']) . '.' . $db->name($field);
-				}
 				$records = $db->fetchAll($db->buildStatement($query, $model), false, $model->alias);
 
 				if ($records !== false && !empty($records)) {
@@ -148,6 +153,9 @@ class CakeTestFixture extends Object {
  * @access public
  */
 	function drop(&$db) {
+		if (empty($this->fields)) {
+			return false;
+		}
 		$this->Schema->_build(array($this->table => $this->fields));
 		return (
 			$db->execute($db->dropSchema($this->Schema), array('log' => false)) !== false
@@ -193,4 +201,3 @@ class CakeTestFixture extends Object {
 		return $return;
 	}
 }
-?>
